Flips the result measurement display
When ran, Bob's received message was printing flipped (i.e. it printed "01" when it received "10"). The issue wasn't in the measurement, because when printing "res.measurements" it correctly assigned the values to the qubits from the message as expected.

This correction flips the bitstring display so it prints the message correctly.

The error might come from a change in how cirq saves measurements from simulators, but I'm not sure.